requirements.txt
pyautogui==0.9.54
keyboard==0.13.5
cx-Freeze==6.15.0
Pillow>=9.5.0 